[Refactor ⚙️] 반응형 웹에 필요한 컴포넌트 분리
