+++ src/pages/ProfilePage/components/ProfileBar.tsx
@@ -1,12 +1,12 @@
 import { Box, Center, Flex, StackDivider, VStack } from "@chakra-ui/react"
 import { TechStackType } from "api-models"
 
-import Introduction from "./Introduction"
 import ProfileCard from "./ProfileCard"
+import ProfileIntroduction from "./ProfileIntroduction"
 import TechStack from "./TechStack"
 
 export interface ProfileBarProps {
-  nickName: string | undefined
+  nickname: string | undefined
   profileImageUrl: string | undefined
   career: string | undefined
   introduction: string | undefined
@@ -16,7 +16,7 @@ export interface ProfileBarProps {
 }
 
 const ProfileBar = ({
-  nickName,
+  nickname,
   profileImageUrl,
   career,
   introduction,
@@ -24,13 +24,10 @@ const ProfileBar = ({
   blogUrl,
   techStacks,
 }: ProfileBarProps) => {
-  console.log(nickName, "여기")
   return (
     <Center
       bg="white"
       w="36rem"
-      //   w="20%"
-      //   top="25rem"
       mt="-10rem"
       border="1px solid darkgrey"
       borderRadius="20px">
@@ -42,13 +39,13 @@ const ProfileBar = ({
           minH="30rem"
           alignItems="center">
           <ProfileCard
-            nickname={nickName}
+            nickname={nickname}
             profileImageUrl={profileImageUrl}
             career={career}
           />
         </Flex>
         <Box>
-          <Introduction
+          <ProfileIntroduction
             aboutMe={introduction}
             githubUrl={githubUrl}
             blogUrl={blogUrl}
